rename *.handler.ts to *.events.ts
